opinions: handle authentication to resources

+++ opinions/src/main/kotlin/com/oxygensend/opinions/context/service/OpinionService.kt
@@ -1,6 +1,8 @@
 package com.oxygensend.opinions.context.service
 
 import com.oxygensend.commons_jdk.PagedListView
+import com.oxygensend.commons_jdk.exception.AccessDeniedException
+import com.oxygensend.commons_jdk.request_context.RequestContext
 import com.oxygensend.opinions.context.command.AddCommentCommand
 import com.oxygensend.opinions.context.command.CreateOpinionCommand
 import com.oxygensend.opinions.context.command.UpdateOpinionCommand
@@ -25,8 +27,10 @@ class OpinionService(
     private val opinionRepository: OpinionRepository,
     private val opinionAggregateRepository: OpinionAggregateRepository,
     private val userRepository: UserRepository,
-    private val applicationEventPublisher: ApplicationEventPublisher
+    private val applicationEventPublisher: ApplicationEventPublisher,
+    private val requestContext: RequestContext
 ) {
+
     fun getOpinions(filter: OpinionsFilter, pageable: Pageable): PagedListView<OpinionView> {
         val paginator = opinionAggregateRepository.findAggregatedOpinions(filter, pageable)
         val data = paginator.map { OpinionView.from(it) }.toList()
@@ -35,6 +39,10 @@ class OpinionService(
 
     fun updateOpinion(opinionId: ObjectId, command: UpdateOpinionCommand) {
         val opinion = opinionRepository.findById(opinionId) ?: throw OpinionNotFoundException()
+
+        if (!requestContext.isUserAuthenticated(opinion.author)) {
+            throw AccessDeniedException()
+        }
 
         updateIfDefined(command.text) {
             opinion.text = command.text.value!!
@@ -49,6 +57,10 @@ class OpinionService(
 
     fun deleteOpinion(opinionId: ObjectId) {
         val opinion = opinionRepository.findById(opinionId) ?: throw OpinionNotFoundException()
+        if (!requestContext.isUserAuthenticated(opinion.author)) {
+            throw AccessDeniedException()
+        }
+
         opinionRepository.delete(opinion)
         publishRecalculationEvent(opinion.receiver)
     }
@@ -56,6 +68,11 @@ class OpinionService(
     fun createOpinion(command: CreateOpinionCommand): OpinionView {
         validateIfCreationIsPossible(command.authorId, command.receiverId)
         val author = userRepository.findById(command.authorId) ?: throw NoSuchUserException("Author with id ${command.authorId} not found")
+
+        if (!requestContext.isUserAuthenticated(command.authorId)) {
+            throw AccessDeniedException()
+        }
+
         val opinion = Opinion(
             id = ObjectId(),
             author = command.authorId,
@@ -73,6 +90,9 @@ class OpinionService(
     fun addLike(opinionId: ObjectId, userId: String) {
         val opinion = opinionRepository.findById(opinionId) ?: throw OpinionNotFoundException()
         userRepository.findById(userId) ?: throw NoSuchUserException("User with id $userId not found")
+        if (!requestContext.isUserAuthenticated(userId)) {
+            throw AccessDeniedException()
+        }
         if (opinion.hasLiked(userId)) {
             throw OpinionLikeException.opinionAlreadyLiked()
         }
@@ -83,6 +103,9 @@ class OpinionService(
     fun addDislike(opinionId: ObjectId, userId: String) {
         val opinion = opinionRepository.findById(opinionId) ?: throw OpinionNotFoundException()
         userRepository.findById(userId) ?: throw NoSuchUserException("User with id $userId not found")
+        if (!requestContext.isUserAuthenticated(userId)) {
+            throw AccessDeniedException()
+        }
         if (!opinion.hasLiked(userId)) {
             throw OpinionLikeException.opinionNotLiked()
         }
@@ -93,6 +116,9 @@ class OpinionService(
     fun addComment(opinionId: ObjectId, command: AddCommentCommand): CommentView {
         val opinion = opinionRepository.findById(opinionId) ?: throw OpinionNotFoundException()
         val author = userRepository.findById(command.author) ?: throw NoSuchUserException("User with id ${command.author} not found")
+        if (!requestContext.isUserAuthenticated(command.author)) {
+            throw AccessDeniedException()
+        }
 
         val comment = Opinion.Comment(ObjectId(), command.author, command.text)
         opinion.addComment(comment)
@@ -104,6 +130,9 @@ class OpinionService(
     fun deleteComment(opinionId: ObjectId, commentId: ObjectId) {
         val opinion = opinionRepository.findById(opinionId) ?: throw OpinionNotFoundException()
         val comment = opinion.findCommentById(commentId) ?: throw NoSuchCommentException("Comment with id $commentId not found")
+        if (!requestContext.isUserAuthenticated(comment.author)) {
+            throw AccessDeniedException()
+        }
         opinion.removeComment(comment)
         opinionRepository.save(opinion)
     }
